% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeDrug2.R
\name{ThreeDrug_test}
\alias{ThreeDrug_test}
\title{Testing the additive effect}
\usage{
ThreeDrug_test(location1, location2, location3, location4, r)
}
\arguments{
\item{location1}{location of Drug 1's dataset on your hard disk}

\item{location2}{location of Drug 2's dataset on your hard disk}

\item{location3}{location of Drug 3's dataset on your hard disk}

\item{location4}{location of combination dataset on your hard disk}

\item{r}{number of replications at each dose-mixture, and usually 3<=r<=8}
}
\value{
F test of additive/synergistic/antagonistic effect under drug combination
}
\description{
This function implements an F test to find out the additive/synergistic/antagonistic effect under each three
drug combination level
}
\examples{
# Try this function with the following example:
SY.test3<-ThreeDrug_test(
location1="C:/Users/58412/Desktop/SRA/Three drugs/data2/Drug1.txt", 
location2="C:/Users/58412/Desktop/SRA/Three drugs/data2/Drug2.txt", 
location3="C:/Users/58412/Desktop/SRA/Three drugs/data2/Drug3.txt", 
location4="C:/Users/58412/Desktop/SRA/Three drugs/data2/Combination.txt", 
r=6)
}
