% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDrug1.R
\name{TwoDrug_identify_design}
\alias{TwoDrug_identify_design}
\title{Identifying the linear/log-linear model of each dose-response curve}
\usage{
TwoDrug_identify_design(location1, location2, location3, c0, r)
}
\arguments{
\item{location1}{location of Drug 1's dataset on your hard disk}

\item{location2}{location of Drug 2's dataset on your hard disk}

\item{location3}{location of UniformDesign on your hard disk}

\item{c0}{the smallest meaningful difference to be detected}

\item{r}{number of replications at each dose-mixture, and usually 3<=r<=8, default=3}
}
\value{
The fitted models and method for two drugs, 
        the number of dose-mixtures for combination, 
        total sample size,
        fitted plots for two drugs,
        dose mixtures for experimental design
}
\description{
This function utilized nls() function to model the HILL model of each relationship and substitute each relationship
with linear or log-linear relationship, it then designs a dose-mixture combination dataset for the study by using Uniform design.
}
\examples{
# Try this function with the following example:
TwoDrug_identify_design1<-TwoDrug_identify_design(
 location1="C:/Users/58412/Desktop/SRA/Two drugs/Drug1.txt",
 location2="C:/Users/58412/Desktop/SRA/Two drugs/Drug2.txt",
 location3="C:/Users/58412/Desktop/SRA/Two drugs/UniformDesign.txt",
 c0=0.01, r=6)
}
