% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDrug2.R
\name{TwoDrug_Test_Analysis}
\alias{TwoDrug_Test_Analysis}
\title{Testing the additive effect and drawing a contour plot}
\usage{
TwoDrug_Test_Analysis(location1, location2, location4, r = 5,
  method = "loglog")
}
\arguments{
\item{location1}{location of Drug 1's dataset on your hard disk}

\item{location2}{location of Drug 2's dataset on your hard disk}

\item{location4}{location of combination dataset on your hard disk}

\item{r}{number of replications at each dose-mixture, and usually 3<=r<=8}

\item{method}{input the method you received in the previous step}
}
\value{
F test of additive/synergistic/antagonistic effect under drug combination,
        a contour plot showing this effect
}
\description{
This function implements an F test to find out the additive/synergistic/antagonistic effect under each two
drug combination level, and draws a contour plot showing this effect
}
\examples{
# Try this function with the following example:
TwoDrug_Test_Analysis <- TwoDrug_Test_Analysis(
 location1="C:/Users/58412/Desktop/SRA/Two drugs/Drug1.txt",
 location2="C:/Users/58412/Desktop/SRA/Two drugs/Drug2.txt",
 location4 = "C:/Users/58412/Desktop/SRA/Two drugs/Combination.txt", 
                                              r=6,method="logli")
}
